feat(jobs): clarify Duration format, add timing to Results page
- Change sub-hour Duration format from `4:00` to `4m 00s` so units are
  unambiguous at a glance
- Extract formatDateTime + formatDuration to src/utils/dateFormat.js so
  both pages share one implementation
- Add Queued / Started / Duration row to the Results page header, matching
  the Jobs table columns
