+++ frontend/src/utils/dateFormat.js
@@ -1,0 +1,19 @@
+export function formatDateTime(dateStr) {
+  if (!dateStr) return '--';
+  const d = new Date(dateStr);
+  return d.toLocaleString(undefined, { month: 'short', day: 'numeric', hour: '2-digit', minute: '2-digit' });
+}
+
+export function formatDuration(startStr, endStr, nowFn) {
+  if (!startStr) return '—';
+  const end = endStr ? new Date(endStr) : new Date((nowFn || Date.now)());
+  const diffSec = Math.max(0, Math.floor((end - new Date(startStr)) / 1000));
+  if (diffSec < 3600) {
+    const m = Math.floor(diffSec / 60);
+    const s = diffSec % 60;
+    return `${m}m ${String(s).padStart(2, '0')}s`;
+  }
+  const h = Math.floor(diffSec / 3600);
+  const m = Math.floor((diffSec % 3600) / 60);
+  return `${h}h ${m}m`;
+}
